frontend: common:  Refactor imports for /common
