+++ frontend/src/components/App/Settings/Settings.tsx
@@ -27,7 +27,9 @@ import { setAppSettings } from '../../../redux/configSlice';
 import { defaultTableRowsPerPageOptions } from '../../../redux/configSlice';
 import { useTypedSelector } from '../../../redux/reducers/reducers';
 import { uiSlice } from '../../../redux/uiSlice';
-import { ActionButton, NameValueTable, SectionBox } from '../../common';
+import ActionButton from '../../common/ActionButton';
+import NameValueTable from '../../common/NameValueTable';
+import SectionBox from '../../common/SectionBox';
 import TimezoneSelect from '../../common/TimezoneSelect';
 import { setTheme, useAppThemes } from '../themeSlice';
 import DrawerModeSettings from './DrawerModeSettings';
