+++ frontend/src/components/App/Settings/SettingsButton.tsx
@@ -18,7 +18,7 @@ import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { getCluster } from '../../../lib/cluster';
 import { createRouteURL } from '../../../lib/router';
-import { ActionButton } from '../../common';
+import ActionButton from '../../common/ActionButton';
 
 export default function SettingsButton(props: { onClickExtra?: () => void }) {
   const { onClickExtra } = props;
