+++ frontend/src/components/App/Settings/SettingsClusters.tsx
@@ -16,7 +16,9 @@
 
 import { useTranslation } from 'react-i18next';
 import { useClustersConf } from '../../../lib/k8s';
-import { Link, SectionBox, SimpleTable } from '../../common';
+import Link from '../../common/Link';
+import SectionBox from '../../common/SectionBox';
+import SimpleTable from '../../common/SimpleTable';
 
 export default function SettingsClusters() {
   const clusterConf = useClustersConf();
