+++ frontend/src/components/storage/VolumeDetails.tsx
@@ -17,7 +17,7 @@
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import PersistentVolume from '../../lib/k8s/persistentVolume';
-import { Link } from '../common';
+import Link from '../common/Link';
 import { DetailsGrid } from '../common/Resource';
 import { StatusLabelByPhase } from './utils';
